rename "dump" to "export"
